refactor: split CalibreData.swift into focused model files (P2/A11)
Decompose the 1307-line Models/CalibreData.swift into ten zero-behavior-change
files and delete the original:

- Models/CalibreCoreModels.swift: CalibreServer, CalibreLibrary, CalibreBook,
  CalibreSyncStatus.
- Models/ReadingPositionModels.swift: BookDeviceReadingPosition,
  BookDeviceReadingPositionHistory, reading statistics.
- Models/CalibreHighlightStyle.swift: BookHighlightStyle (only split file that
  imports UIKit).
- Models/CalibreTasks.swift: network/metadata task structs.
- Models/CalibrePayloadModels.swift: Codable API entry/result payloads.
- Models/CalibreSyncModels.swift: custom columns, category keys, probe/sync
  requests, CalibreCdbCmdListResult.
- Models/CalibrePluginModels.swift: DSReader Helper / Count Pages / Goodreads
  Sync prefs and CalibreDSReaderHelperConfiguration.
- Network/CalibreActivityModels.swift: CalibreActivity/Start/Finish.
- Models/CalibreServerConfigProvider.swift: CalibreServerConfigProvider protocol.
- Models/Array+Chunks.swift: generic Array.chunks(size:) helper.

All new files are registered in both app and Catalyst targets. Added
YetAnotherEBookReaderTests/CalibreDataSplitTests (20 cases) covering
CalibreServer/CalibreLibrary identity and hashing, CalibreBook.inShelfId,
BookHighlightStyle class/color mapping, plugin preference Codable decoding,
custom-column snake_case CodingKeys, Array.chunks, and the CalibreActivity
class hierarchy.

Verified with xcodebuild build and xcodebuild test on iPhone 17 simulator:
87 unit tests + 1 UI test green (was 67 + 1). The Mac Catalyst build remains
blocked by a pre-existing SPM package product resolution issue
(R2Navigator, GCDWebServer, R2Shared, R2Streamer) unrelated to this split.

Updated AGENTS.md Architecture Map and handoff context, and marked the
p2-calibredata-split-plan action items complete.
