+++ .agents/memory-bank/activeContext.md
@@ -1,9 +1,11 @@
 # Active Context
 
 ## Current Focus
-The primary focus is the Modernization of the EBook Reader Architecture. We have completed milestones P0-1a (`CalibreServerManager`), P0-1b (`CalibreLibraryManager`), P0-1d (`CalibreBookManager`), P0-1e (`ReadingSessionManager`), P0-3a (`BookRepository`), P0-3b (`ServerRepository` and `LibraryRepository`), P0-3c (`AnnotationRepository`), P1a-A06 (`ReaderEngineDelegate` position consolidation), P1b-A03 (Dismantling YabrPDFViewController), P1d-A04 (`CalibreServerService` refactor), P1e (V2 migration / CalibreBrowser V1 removal), and P1f (`MainViewModel` and MVVM Modernization) to modularize the application state, decouple network/management operations, and separate concerns. The latest verified test run passes 67 unit tests and 1 UI test.
+The primary focus is the Modernization of the EBook Reader Architecture. We have completed milestones P0-1a (`CalibreServerManager`), P0-1b (`CalibreLibraryManager`), P0-1d (`CalibreBookManager`), P0-1e (`ReadingSessionManager`), P0-3a (`BookRepository`), P0-3b (`ServerRepository` and `LibraryRepository`), P0-3c (`AnnotationRepository`), P1a-A06 (`ReaderEngineDelegate` position consolidation), P1b-A03 (Dismantling YabrPDFViewController), P1d-A04 (`CalibreServerService` refactor), P1e (V2 migration / CalibreBrowser V1 removal), P1f (`MainViewModel` and MVVM Modernization), and P2/A11 (`CalibreData.swift` split) to modularize the application state, decouple network/management operations, and separate concerns. The latest verified test run passes 87 unit tests and 1 UI test.
 
 ## Recent Changes & Decisions
+- **P2/A11 CalibreData Split Implemented (2026-06-18):** Decomposed the 1307-line `Models/CalibreData.swift` into ten focused files as a zero-behavior-change move: `Models/CalibreCoreModels.swift` (`CalibreServer`, `CalibreLibrary`, `CalibreBook`, `CalibreSyncStatus`), `Models/ReadingPositionModels.swift` (`BookDeviceReadingPosition`, `BookDeviceReadingPositionHistory`, statistics), `Models/CalibreHighlightStyle.swift` (`BookHighlightStyle` — only file in the split that imports UIKit), `Models/CalibreTasks.swift` (network/metadata task structs), `Models/CalibrePayloadModels.swift` (Codable API entry/result payloads), `Models/CalibreSyncModels.swift` (custom columns, category keys, probe/sync requests, cdb cmd list result), `Models/CalibrePluginModels.swift` (DSReader Helper / Count Pages / Goodreads Sync prefs + configuration), `Network/CalibreActivityModels.swift` (`CalibreActivity`/`Start`/`Finish`), `Models/CalibreServerConfigProvider.swift` (the protocol), and `Models/Array+Chunks.swift`. Deleted the original `CalibreData.swift`. All new files registered in both app and Catalyst targets via the Xcode project. Added `CalibreDataSplitTests` (20 cases) covering `CalibreServer`/`CalibreLibrary` identity and hashing, `CalibreBook.inShelfId`, `BookHighlightStyle` class/color mapping, plugin preference Codable decoding, custom-column snake_case CodingKeys, `Array.chunks`, and `CalibreActivity` class hierarchy. Full `xcodebuild test` green: 87 unit tests + 1 UI test. Catalyst build still blocked by a pre-existing SPM package product resolution issue (`R2Navigator`, `GCDWebServer`, `R2Shared`, `R2Streamer`) that is unrelated to the split.
+- **P2/A11 CalibreData Split Plan Captured (2026-06-18):** Added `.agents/memory-bank/p2-calibredata-split-plan.md` after analyzing the 1307-line `Models/CalibreData.swift`. The recommended implementation is a zero-behavior split into focused files for core Calibre domain models, reading-position models, highlight style, network task DTOs, Codable payload models, sync/category models, plugin preference models, activity models, `CalibreServerConfigProvider`, and `Array.chunks(size:)`. Preserve type names, Codable keys, hashing/identity semantics, and Xcode target membership; defer semantic cleanup of `CalibreBook`, plugin fallback lookups, and legacy comments to later work.
 - **P1f / MVVM Modernization Completed (2026-06-18):** Enforced clean MVVM pattern on the highest traffic views (`MainView`, `SettingsView`, `SupportInfoView`, `RecentShelfController`, and `SectionShelfController`) to resolve monolithic state trapping in SwiftUI view bodies and UIKit controllers. Introduced `MainViewModel`, `SettingsViewModel`, `SupportInfoViewModel`, `RecentShelfViewModel`, and `SectionShelfViewModel` with thorough `@MainActor`-isolated unit tests covering actions, state, and mutations. Consolidated LibraryInfo child-view commands into the existing V2 ViewModels. Verified the entire project compiles successfully and runs with 67 unit tests + 1 UI test green.
 - **P1f/A08 Plan Captured (2026-06-18):** Added `.agents/memory-bank/p1f-main-viewmodel-plan.md` with the proposed plan for introducing ViewModels around remaining high-traffic views. The recommended order is `MainViewModel`, `SettingsViewModel`, LibraryInfo child-view command extraction, `SupportInfoViewModel`, then shelf ViewModels for `RecentShelfController` and `SectionShelfController`. Reader engine adapters remain out of first-pass scope unless acceptance criteria expand.
 - **P1e Commit History Corrected (2026-06-17):** Rewrote the pushed `HEAD` commit message for the P1e migration so the history matches the actual change set. The corrected commit is `b15c0c1` (`refactor: complete V2 search/category migration and remove CalibreBrowser V1`) and replaces the stale PDF-oriented message previously attached to the same diff.
@@ -86,6 +88,7 @@ The primary focus is the Modernization of the EBook Reader Architecture. We have
 - [x] 43. Fixed `RealmReadingPositionRepository` cross-thread Realm reuse and added `ReadingPositionRepositoryThreadingTests`; full test suite now passes again.
 - [x] 44. Execute P1d-A04: refactor `CalibreServerService` by introducing `CalibreAPIError`, extracting a shared request/error-mapping layer, and splitting endpoint logic by domain while preserving facade compatibility. Added `CalibreServerServiceTests` and verified with `xcodebuild test` (51 unit tests + 1 UI test).
 - [x] 45. Execute P1e: promote V2 search/category services to `ModelData`, remove `CalibreLibrarySearchManager` / `CalibreBrowser.swift`, delete V1 unified category Realm models, add migration cleanup for schema `139`, and verify the direct V2 dependency graph with `V2MigrationDependencyTests`. Full validation now passes with 55 unit tests + 1 UI test.
+- [x] 46. Execute P2/A11: split `Models/CalibreData.swift` (1307 lines) into ten focused model/payload/task/plugin/activity files, delete the original file, register all new files in both app and Catalyst targets, and add `CalibreDataSplitTests` (20 cases) covering identity/hashing, `inShelfId`, `BookHighlightStyle` mapping, plugin preference Codable decoding, custom-column CodingKeys, `Array.chunks`, and `CalibreActivity` hierarchy. Full validation green: 87 unit tests + 1 UI test.
 
 ## Completed Refactoring Milestones
 - [x] Extract `CalibreServerManager` out of `ModelData` (Milestone P0-1a).
@@ -107,6 +110,7 @@ The primary focus is the Modernization of the EBook Reader Architecture. We have
 - [x] Correct FolioReader font size mapping discrepancy and align with unified preferences.
 - [x] Mitigate Realm database hangs/UI freezes when opening BookDetailView (Fixes 1-5).
 - [x] Remove books property from CalibreLibrarySearchValueObject and dynamically resolve books list.
+- [x] Split `Models/CalibreData.swift` into ten focused files (Milestone P2/A11) and add boundary tests for identity, Codable, and style mappings.
 
 ## Active Constraints
 - **Do NOT** introduce CocoaPods or modify workspace files; the project relies entirely on Swift Package Manager.
