+++ YetAnotherEBookReader/Models/Array+Chunks.swift
@@ -1,0 +1,17 @@
+//
+//  Array+Chunks.swift
+//  YetAnotherEBookReader
+//
+//  Split from CalibreData.swift on 2026/6/18.
+//  Zero-behavior-change move: generic Array.chunks(size:) helper.
+//
+
+import Foundation
+
+extension Array {
+    func chunks(size: Int) -> [[Element]] {
+        return stride(from: 0, to: count, by: size).map {
+            Array(self[$0 ..< Swift.min($0 + size, count)])
+        }
+    }
+}
